requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
httpx==0.27.0
feedparser==6.0.11
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
